refactor: extract async status projection
