+++ test/unit/async-status-projection.test.ts
@@ -1,0 +1,86 @@
+import assert from "node:assert/strict";
+import { describe, it } from "node:test";
+import type { AsyncJobState } from "../../src/shared/types.ts";
+import {
+	ASYNC_STATUS_SNAPSHOT_KIND,
+	ASYNC_STATUS_SNAPSHOT_VERSION,
+	projectAsyncWorkflowRows,
+	projectAsyncStatusSnapshot,
+} from "../../src/runs/shared/async-status-projection.ts";
+
+function job(input: Partial<AsyncJobState> & Pick<AsyncJobState, "asyncId" | "status">): AsyncJobState {
+	return {
+		asyncDir: `/tmp/${input.asyncId}`,
+		...input,
+	} as AsyncJobState;
+}
+
+describe("async status projection", () => {
+	it("projects already-loaded jobs in deterministic newest-first order", () => {
+		const snapshot = projectAsyncStatusSnapshot([
+			job({ asyncId: "older", status: "complete", agents: ["reviewer"], updatedAt: 10 }),
+			job({
+				asyncId: "newer",
+				status: "running",
+				mode: "workflow",
+				agents: ["worker"],
+				updatedAt: 20,
+				steps: [{ agent: "worker", status: "pending" }],
+			}),
+		], { generatedAt: 30 });
+
+		assert.equal(snapshot.kind, ASYNC_STATUS_SNAPSHOT_KIND);
+		assert.equal(snapshot.version, ASYNC_STATUS_SNAPSHOT_VERSION);
+		assert.equal(snapshot.generatedAt, 30);
+		assert.deepEqual(snapshot.runs.map(({ id, kind, state }) => ({ id, kind, state })), [
+			{ id: "newer", kind: "workflow", state: "running" },
+			{ id: "older", kind: "subagent", state: "complete" },
+		]);
+		assert.equal(snapshot.runs[0]?.children?.[0]?.state, "queued");
+	});
+
+	it("preserves partial needs-attention status while excluding private evidence", () => {
+		const snapshot = projectAsyncStatusSnapshot([job({
+			asyncId: "partial-run",
+			status: "partial",
+			agents: ["writer"],
+			activityState: "needs_attention",
+			steps: [{
+				agent: "writer",
+				status: "partial",
+				activityState: "needs_attention",
+				error: "Required file-only output was not produced: /private/report.md",
+				effects: { fileMutation: true },
+			}],
+		})], { generatedAt: 1 });
+
+		assert.equal(snapshot.runs[0]?.state, "partial");
+		assert.equal(snapshot.runs[0]?.activity?.state, "needs_attention");
+		assert.equal(snapshot.runs[0]?.children?.[0]?.state, "partial");
+		assert.equal(snapshot.runs[0]?.children?.[0]?.activity?.state, "needs_attention");
+		const serialized = JSON.stringify(snapshot);
+		assert.doesNotMatch(serialized, /private\/report|fileMutation|Required file-only output/);
+	});
+
+	it("projects compact Fleet workflow rows without applying UI bounds", () => {
+		const rows = projectAsyncWorkflowRows([{
+			agent: "reviewer",
+			workflowKey: "review",
+			label: "Fresh review",
+			phase: "quality",
+			status: "partial",
+			activityState: "needs_attention",
+			startedAt: 10,
+			tokens: { input: 20, output: 5, total: 25, window: 18 },
+		}]);
+
+		assert.deepEqual(rows, [{
+			name: "quality: review · Fresh review (reviewer)",
+			state: "partial",
+			activity: "needs attention",
+			startedAt: 10,
+			tokens: 25,
+			window: 18,
+		}]);
+	});
+});
